requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.2
kafka-python==2.0.2
requests==2.31.0